requirements.txt
# requirements.txt
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
tensorflow==2.5.0
matplotlib==3.4.2
seaborn==0.11.1
pyyaml==5.4.1